+++ packages/vue/src/hover-card/hover-card-arrow-tip.tsx
@@ -1,11 +1,11 @@
 import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import { getValidChildren, type ComponentWithProps } from '../utils'
+import { getValidChildren } from '../utils'
 import { useHoverCardContext } from './hover-card-context'
 
 export type HoverCardArrowTipProps = HTMLArkProps<'div'>
 
-export const HoverCardArrowTip: ComponentWithProps<HoverCardArrowTipProps> = defineComponent({
+export const HoverCardArrowTip = defineComponent({
   name: 'HoverCardArrowTip',
   setup(_, { slots, attrs }) {
     const api = useHoverCardContext()