+++ packages/vue/src/hover-card/hover-card-context.ts
@@ -1,9 +1,7 @@
-import { type connect } from '@zag-js/hover-card'
-import { type ComputedRef } from 'vue'
 import { createContext } from '../context'
 import { type UseHoverCardReturn } from './use-hover-card'
 
-export const [HoverCardProvider, useHoverCardContext] =
-  createContext<ComputedRef<ReturnType<typeof connect>>>('HoverCardContext')
-
 export type HoverCardContext = UseHoverCardReturn
+
+export const [HoverCardProvider, useHoverCardContext] =
+  createContext<HoverCardContext>('HoverCardContext')